+++ rust/src/commands/mod.rs
@@ -7,7 +7,6 @@ mod create_token_account;
 mod mint_tokens;
 mod request_airdrop;
 mod send_tokens;
-mod test;
 
 #[derive(Parser)]
 #[command(name = "solana-quiz")]
@@ -38,7 +37,6 @@ pub enum Commands {
         amount: u64,
     },
     CreateMetadataAccounts {},
-    Test {},
 }
 
 /// Runs the CLI application by parsing the user command and executing
@@ -64,9 +62,6 @@ pub async fn run() -> Result<()> {
         }
         Commands::CreateMetadataAccounts {} => {
             create_metadata_accounts::run().await?;
-        }
-        Commands::Test {} => {
-            test::run().await?;
         }
     }